// dashboard/src/api/models/quoteStatusEnum.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `draft` - Draft
 * `open` - Open
 * `accepted` - Accepted
 * `canceled` - Canceled
 */
export type QuoteStatusEnum =
  (typeof QuoteStatusEnum)[keyof typeof QuoteStatusEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const QuoteStatusEnum = {
  draft: "draft",
  open: "open",
  accepted: "accepted",
  canceled: "canceled",
} as const;